event source bug fix
eventSourceRef.current = null was missing in the return of the useEffect. This would lead to a problem when leaving and re-entering a course, as no new event source would be created (since ref.current !== null), but the EventSource stored in the ref is closed and therefore no longer reacts to messages from the backend.