fixes unable to retreive data for not currently displaye files bug
This change fixes another bug where unsaved content for files not actively visible in the editor was lost after a submission, which caused them revert to their template content.

The Monaco Editor uses the path prop to create and manage unique Model instances. The original getPath function included the submissionId, which was changing on every run/submission:

```typescript
const getPath = (id: number) => `${id}/${user.email}/${submissionId}`
```

When a run/submission was made, submissionId updated and changed the required model path for all files. In that case the Monaco Editor destroyed the old model because it was no longer linked to an active component path.
When we later called editor.getContent(path) for an inactive file, no model existed at the new path, which lead the function to return undefined and wrongly fall back to file.template, overwriting the content from the actual submission.

I checked that the submissionId is only necessary for data fetching and is not required for Monaco's model identity, since we manually set the content of the editor when restoring a specific  implementation based on the submission's content.